# perf(compiler-cli): optimize cycle detection using a persistent cache