--- a/packages/compiler-cli/src/ngtsc/cycles/test/imports_spec.ts
+++ b/packages/compiler-cli/src/ngtsc/cycles/test/imports_spec.ts
@@ -28,34 +28,6 @@ runInEachFileSystem(() => {
       });
     });
 
-    describe('transitiveImportsOf()', () => {
-      it('should calculate transitive imports of a simple program', () => {
-        const {program, graph} = makeImportGraph('a:b;b:c;c');
-        const a = getSourceFileOrError(program, (_('/a.ts')));
-        const b = getSourceFileOrError(program, (_('/b.ts')));
-        const c = getSourceFileOrError(program, (_('/c.ts')));
-        expect(importsToString(graph.transitiveImportsOf(a))).toBe('a,b,c');
-      });
-
-      it('should calculate transitive imports in a more complex program (with a cycle)', () => {
-        const {program, graph} = makeImportGraph('a:*b,*c;b:*e,*f;c:*g,*h;e:f;f;g:e;h:g');
-        const c = getSourceFileOrError(program, (_('/c.ts')));
-        expect(importsToString(graph.transitiveImportsOf(c))).toBe('c,e,f,g,h');
-      });
-
-      it('should reflect the addition of a synthetic import', () => {
-        const {program, graph} = makeImportGraph('a:b,c,d;b;c;d:b');
-        const b = getSourceFileOrError(program, (_('/b.ts')));
-        const c = getSourceFileOrError(program, (_('/c.ts')));
-        const d = getSourceFileOrError(program, (_('/d.ts')));
-        expect(importsToString(graph.importsOf(b))).toEqual('');
-        expect(importsToString(graph.transitiveImportsOf(d))).toEqual('b,d');
-        graph.addSyntheticImport(b, c);
-        expect(importsToString(graph.importsOf(b))).toEqual('c');
-        expect(importsToString(graph.transitiveImportsOf(d))).toEqual('b,c,d');
-      });
-    });
-
     describe('findPath()', () => {
       it('should be able to compute the path between two source files if there is a cycle', () => {
         const {program, graph} = makeImportGraph('a:*b,*c;b:*e,*f;c:*g,*h;e:f;f;g:e;h:g');