@JoostK 2021-07-13
I am seeing noticeable improvements in several different projects with this change; cycle detection timings going from ~500ms to ~200ms. Since this runs on each incremental rebuild, that is a nice win (overall resolve phase time reduces from ~800ms to ~500ms)